Replaced index with new home page
